# backend/openapi.json
{
  "components": {
    "schemas": {
      "AcquisitionJobListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AcquisitionJobResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "AcquisitionJobListResponse",
        "type": "object"
      },
      "AcquisitionJobResponse": {
        "properties": {
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "byte_count": {
            "title": "Byte Count",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duplicate_count": {
            "title": "Duplicate Count",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "filtered_count": {
            "title": "Filtered Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "new_count": {
            "title": "New Count",
            "type": "integer"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "source_slug": {
            "title": "Source Slug",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "unchanged_count": {
            "title": "Unchanged Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "source_id",
          "source_slug",
          "status",
          "outcome",
          "error_code",
          "attempt_count",
          "new_count",
          "unchanged_count",
          "duplicate_count",
          "filtered_count",
          "byte_count",
          "created_at",
          "started_at",
          "completed_at"
        ],
        "title": "AcquisitionJobResponse",
        "type": "object"
      },
      "AcquisitionRunResponse": {
        "properties": {
          "acquisition_version": {
            "title": "Acquisition Version",
            "type": "string"
          },
          "business_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Date"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "content_slot": {
            "anyOf": [
              {
                "enum": [
                  "morning",
                  "noon",
                  "evening"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Slot"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duplicate_count": {
            "title": "Duplicate Count",
            "type": "integer"
          },
          "failed_jobs": {
            "title": "Failed Jobs",
            "type": "integer"
          },
          "filtered_count": {
            "title": "Filtered Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "new_count": {
            "title": "New Count",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          },
          "succeeded_jobs": {
            "title": "Succeeded Jobs",
            "type": "integer"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "total_jobs": {
            "title": "Total Jobs",
            "type": "integer"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          },
          "unchanged_count": {
            "title": "Unchanged Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "trigger",
          "business_date",
          "timezone",
          "acquisition_version",
          "content_slot",
          "status",
          "total_jobs",
          "succeeded_jobs",
          "failed_jobs",
          "new_count",
          "unchanged_count",
          "duplicate_count",
          "filtered_count",
          "created_at",
          "started_at",
          "completed_at",
          "status_url"
        ],
        "title": "AcquisitionRunResponse",
        "type": "object"
      },
...
      "CreateAcquisitionRunRequest": {
        "properties": {
          "business_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional isolated business date used by local preview runs.",
            "title": "Business Date"
          },
          "source_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "maxItems": 9,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ids"
          }
        },
        "title": "CreateAcquisitionRunRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Internal API for authoritative-source acquisition, factual governance, and auditable event organization, deterministic daily topic selection, and private brand knowledge ingestion with internal copy-generation context retrieval. It does not expose public brand search, publishing, or arbitrary URL-fetch capabilities.",
    "title": "Edu AI Lead Agent API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/acquisition-runs": {
      "post": {
        "operationId": "create_acquisition_run_api_v1_acquisition_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 8,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAcquisitionRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcquisitionRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Acquisition Run",
        "tags": [
          "acquisition"
        ]
      }
    },
    "/api/v1/acquisition-runs/{run_id}": {
      "get": {
        "operationId": "get_acquisition_run_api_v1_acquisition_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcquisitionRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Acquisition Run",
        "tags": [
          "acquisition"
        ]
      }
    },
    "/api/v1/acquisition-runs/{run_id}/jobs": {
      "get": {
        "operationId": "get_acquisition_run_jobs_api_v1_acquisition_runs__run_id__jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcquisitionJobListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Acquisition Run Jobs",
        "tags": [
          "acquisition"
        ]
      }
    },
...
  }
}